ai-service/requirements.txt
fastapi
uvicorn
torch
torchvision
transformers
pillow
accelerate
sentencepiece
protobuf
tiktoken
